added new CalculateSolarForce by mrsolarsail
I wanted to let you know that the solar sail module has a couple errors
in the force model. If the back side points towards the sun, so does the
normal vector and force vector. Also, the force should be proportional
to the cosine-squared of the angle to the sun, instead of just the
cosine. One cosine term comes from the sail intercepting less light as
it tilts. The second is because the light glances off the sail at
steeper angles, contributing less to thrust.